26/2 15:19

+++ src/main/java/israela/milestone2/User.java
@@ -20,7 +20,7 @@ public class User {
         this.isAdmin = false;
     }
 
-    public double getId() {
+    public Long getId() {
         return id;
     }
